Projection class has been moved to its own file. All (or most) projection specific drawing has been moved to Projection.  Phase lead and lag angle is show on the outside of the vector sweep circle.  Helper function in Projection class allows drawing text in a circular path.
More control over how ordinate lines and captions are shown.  Setting initial values of checkboxes and spinboxes is done using a new method where a mapping is established between the GUI widget and its backing variable using a vector of pairs.  Then a for loop is used to actually set the value into the GUI widget.